+++ setup.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+
 try:
     from setuptools import setup
 except ImportError:
